CSS modification + posts page + weird button
Have performed some modification on the CSS of the website as I found the appearance a little too simple in its whole ( and for my liking... ).
Also added a page where we can send messages anonymously. Said feature isn't included yet, but work is underway to create it...
Added a button for no reason.